requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
numpy
Pillow
scipy
